fix: Added missing context argument to __onLiveSync during HMR refresh.
